+++ src/VisualStudio/IntegrationTest/TestUtilities/LightBulbHelper.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.Language.Intellisense;
 using Microsoft.VisualStudio.Text.Editor;
@@ -11,11 +12,11 @@ namespace Microsoft.VisualStudio.IntegrationTest.Utilities
 {
     public static class LightBulbHelper
     {
-        public static Task<bool> WaitForLightBulbSessionAsync(ILightBulbBroker broker, IWpfTextView view)
+        public static async Task<bool> WaitForLightBulbSessionAsync(ILightBulbBroker broker, IWpfTextView view)
         {
             var startTime = DateTimeOffset.Now;
 
-            return Helper.RetryAsync(async () =>
+            var active = await Helper.RetryAsync(async () =>
             {
                 if (broker.IsLightBulbSessionActive(view))
                 {
@@ -33,6 +34,42 @@ namespace Microsoft.VisualStudio.IntegrationTest.Utilities
 
                 return broker.IsLightBulbSessionActive(view);
             }, TimeSpan.Zero);
+
+            if (!active)
+                return false;
+
+            await WaitForItemsAsync(broker, view);
+            return true;
+        }
+
+        public static async Task<IEnumerable<SuggestedActionSet>> WaitForItemsAsync(ILightBulbBroker broker, IWpfTextView view)
+        {
+            var activeSession = broker.GetSession(view);
+            if (activeSession == null)
+            {
+                var bufferType = view.TextBuffer.ContentType.DisplayName;
+                throw new InvalidOperationException(string.Format("No expanded light bulb session found after View.ShowSmartTag.  Buffer content type={0}", bufferType));
+            }
+
+            var start = DateTime.Now;
+            IEnumerable<SuggestedActionSet> actionSets = Array.Empty<SuggestedActionSet>();
+            while (DateTime.Now - start < Helper.HangMitigatingTimeout)
+            {
+                var status = activeSession.TryGetSuggestedActionSets(out actionSets);
+                if (status is not QuerySuggestedActionCompletionStatus.Completed and
+                              not QuerySuggestedActionCompletionStatus.Canceled)
+                {
+                    await Task.Delay(TimeSpan.FromSeconds(1));
+                    continue;
+                }
+
+                if (status != QuerySuggestedActionCompletionStatus.Completed)
+                    actionSets = Array.Empty<SuggestedActionSet>();
+
+                break;
+            }
+
+            return actionSets;
         }
     }
 }
